requirements.txt
torch~=2.2.2
transformers~=4.39.2
tqdm~=4.66.2
langdetect~=1.0.9
pandas~=2.2.1
mysql~=0.0.3
mysql-connector-python~=8.3.0
openai~=1.16.2
SQLAlchemy~=2.0.29
python-dotenv~=1.0.1
accelerate~=0.29.3